requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
torch
transformers
pandas
scikit-learn
passlib
PyJWT
jose